Refactor redux slice
