+++ src/slice.js
@@ -1,6 +1,6 @@
 import { createSlice } from '@reduxjs/toolkit';
 
-import { getActivityDescription, getGoalNumber } from './utils';
+import { getCalories } from './utils';
 
 const { actions, reducer } = createSlice({
   name: 'application',
@@ -14,7 +14,7 @@ const { actions, reducer } = createSlice({
         level: 1,
         description: '운동 거의 안함, 주로 앉아서 앉아서 일하거나 재택 근무로 활동량 매우 적음',
       },
-      goalNubmer: 0,
+      goalNumber: 0,
     },
     calories: {
       bmr: null,
@@ -24,29 +24,6 @@ const { actions, reducer } = createSlice({
   },
   reducers: {
     changeBodyStats(state, { payload: { name, value } }) {
-      if (name === 'activity') {
-        return {
-          ...state,
-          bodyStats: {
-            ...state.bodyStats,
-            activity: {
-              level: value,
-              description: getActivityDescription(value),
-            },
-          },
-        };
-      }
-
-      if (name === 'goalNubmer') {
-        return {
-          ...state,
-          bodyStats: {
-            ...state.bodyStats,
-            goalNubmer: getGoalNumber(value),
-          },
-        };
-      }
-
       return {
         ...state,
         bodyStats: {
@@ -83,7 +60,7 @@ export function calculateCalories() {
         height,
         weight,
         activity,
-        goalNubmer,
+        goalNumber,
       },
     } = getState();
 
@@ -91,18 +68,13 @@ export function calculateCalories() {
       return;
     }
 
-    const correction = gender === 'male' ? 5 : -161;
-    const equation = (10 * weight) + (6.25 * height) - (5 * age);
-    const bmr = Math.round(equation + correction);
+    const { level: activityLevel } = activity;
 
-    const activityNumbers = [1.2, 1.375, 1.55, 1.725, 1.9];
-    const { level } = activity;
+    const calories = getCalories({
+      gender, weight, height, age, activityLevel, goalNumber,
+    });
 
-    const tdee = Math.round(bmr * activityNumbers[level - 1]);
-
-    const result = Math.round(tdee * goalNubmer);
-
-    dispatch(setCalories({ bmr, tdee, result }));
+    dispatch(setCalories(calories));
   };
 }
