+++ src/slice.test.js
@@ -25,7 +25,7 @@ describe('reducer', () => {
             level: 1,
             description: '운동 거의 안함',
           },
-          goalNubmer: 0,
+          goalNumber: 0,
         },
       };
 
@@ -74,7 +74,7 @@ describe('actions', () => {
             level: 1,
             description: '운동 거의 안함',
           },
-          goalNubmer: 1,
+          goalNumber: 1,
         },
       });
     });
